continuing implementation of canopy model
